Add iexamine actors, replace cardreader actor
Terrain/furniture examination has long need a system where, instead of a
hardcoded action, the changes are performed based on data specified in
the terrain entry. This adds that system.

Add a virtual iexamine_actor, from which subclasses inherit, depending
on the type. This is in large part based on the iuse_actor
infrastructure.
This contains a string informing what type the actor is, for querying
whether a terrain or furniture has this actor and for determining what
subactor to load.

As an example here, the carder iexamine function is converted to be an
examine actor.