Convert ter and furn _id constants to str versions
After #45099, it should now be acceptable to use string_ids for these
instead.

But past that, why convert them? Because when constructing the int_id
versions of these when no corresponding ter/furn entry is loaded, it
gives an error. This is correct, BUT these are always constructed - no
matter if you are actually using the 'dda' content to which they belong
or not. Converting them to string ids pushes this error down the road to
where they are actually used - which, when the content associated with
them is not loaded, should be nowhere.

Convert them, convert map::ter and map::furn to return string ids (less
changes result from this), and clean things up. I only had to add new
string->int id conversions, because fixing those remaining few would
have massively ballooned this.