Add tests for crafting failure chance
Add a number of benchmark recipes which need only a 2x4 for
tools/ingredients, and test that their actual failure rates match the
calculated one and pre-provided ones at various levels of skill and
proficiency.

Also add a synthetic test to output the calculated chances for every
recipe in the game.